+++ disasm.go
@@ -49,9 +49,10 @@ func (p *Prog) disasmInstr(offset int) int {
 
 	case opBIND:
 		return bindInstr(p.output, instr, p, offset)
-
 	case opBINDNB:
 		return bindnbInstr(p.output, instr, p, offset)
+	case opBINDNBS:
+		return bindnbsInstr(p.output, instr, p, offset)
 
 	default:
 		fmt.Fprintln(p.output, "unknown opcode", instr)
@@ -115,4 +116,27 @@ func bindnbInstr(w io.Writer, o opcode, p *Prog, offset int) int {
 		o, idx1, p.constants[idx1], idx2, p.constants[idx2], arg,
 	)
 	return offset + 1 + n + m + 1
+}
+
+func bindnbsInstr(w io.Writer, o opcode, p *Prog, offset int) int {
+	tidx, n := uvarintFromBytes(p.code[offset+1:])
+	ncnt, m := uvarintFromBytes(p.code[offset+1+n:])
+	names := make([]struct {
+		idx uint64
+		v   value
+	}, ncnt)
+	var k int // cumulative size of name consts
+	for i := uint64(0); i < ncnt; i++ {
+		idx, j := uvarintFromBytes(p.code[offset+1+n+m+k:])
+		names[i].idx, names[i].v = idx, p.constants[idx]
+		k += j
+	}
+	arg := p.code[offset+1+n+m+k]
+
+	fmt.Fprintf(w, "%-10s %4d '%v'", o, tidx, p.constants[tidx])
+	for _, name := range names {
+		fmt.Fprintf(w, "\t%4d '%v'", name.idx, name.v)
+	}
+	fmt.Fprintf(w, "\t0x%2X\n", arg)
+	return offset + 1 + n + m + k + 1
 }